Renamed `builders` submodule to `strategies`
