+++ src/histogram/histograms.rs
@@ -99,7 +99,7 @@ pub trait HistogramExt<A, S>
     /// extern crate noisy_float;
     /// use ndarray_stats::HistogramExt;
     /// use ndarray_stats::histogram::{Histogram, Grid, GridBuilder};
-    /// use ndarray_stats::histogram::builders::Sqrt;
+    /// use ndarray_stats::histogram::strategies::Sqrt;
     /// use noisy_float::types::{N64, n64};
     ///
     /// # fn main() {
