+++ src/histogram/strategies.rs
